+++ NEWS.md
@@ -5,6 +5,9 @@
 * The default renderer used for the RStudio graphics device can now be customized. (#2142)
 * The AGG renderer (as provided by the ragg package) is now a supported backend. (#6539)
 
+### RStudio Server
+
+* The font used in the editor and console can now be customized. (#2534)
 
 ### Bugfixes